refactor(General): :recycle: Move prometheus config in config folder
