+++ src/core/components/HighlightValueCell/HighightValueCell.interfaces.ts
