+++ src/pages/Sites/__tests__/Site.spec.tsx
@@ -4,7 +4,7 @@ import * as router from 'react-router';
 
 import { ProcessResponse, SiteResponse } from '@API/REST.interfaces';
 import { getTestsIds } from '@config/testIds.config';
-import { Wrapper } from '@core/Wrapper';
+import { Wrapper } from '@core/components/Wrapper';
 import processesData from '@mocks/data/PROCESSES.json';
 import sitesData from '@mocks/data/SITES.json';
 import { MockApi, MockApiPaths, loadMockServer } from '@mocks/server';
